@@ -75,8 +75,9 @@
 	"dependencies": {
 		"lucide-react": "^0.477.0",
 		"prop-types": "^15.6.1",
-		"react": "^16.8.4",
-		"react-dom": "^16.8.4",
+		"react": "^19.0.0",
+		"react-calendar": "^5.1.0",
+		"react-dom": "^19.0.0",
 		"react-polyfills": "0.0.1",
 		"react-router-dom": "^6.3.0"
 	}
