@@ -1,6 +1,6 @@
 //import react into the bundle
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 
 //include your index.scss file into the bundle
 import "../styles/index.css";
@@ -9,4 +9,5 @@ import "../styles/index.css";
 import Layout from "./layout";
 
 //render your react application
-ReactDOM.render(<Layout />, document.querySelector("#app"));
+const root = ReactDOM.createRoot(document.querySelector("#app"));
+root.render(<Layout />);
